Create database_helper to help us manage the percona/mysql connection in tests and disabling/enabling the rails advisory patch
